chore: Add Gulp to automate and enhance development
Gulp replaces Grunt as a build and automation tool, assisting on the generation and optimization of responsive images and the optimization of CSS (using PostCSS and paving the way for a future implementation of Sass) and JavaScript (using Browserify + Babelify + Uglify).

New folder structure has been implemented to make use of Gulp's automation features, using a src directory as the development directory, with a .tmp directory as a reference and an intermediary between the src and the dist directories, from which the app will be served.

Gulp assists in watching files for changes during development, with BrowserSync triggering the needed development tasks and synchronizing the changes with the browser.

Minor changes to the app's scripts to address changes introduced by Gulp and its plugins and minor bug addressed in styles.css which caused abnormal sizing of image in restaurant cards when viewed on smaller screens.